moved PerfMetric to own module, using it in TreeProcessor
